fix: apply storybook-dark-mode to icon, icon button
