+++ docs/src/components/Icon/Icon.stories.mdx
@@ -1,5 +1,5 @@
-import {Meta, Story, Canvas} from '@storybook/addon-docs/blocks';
-import {Default} from './index';
+import {Meta} from '@storybook/addon-docs';
+import Icon from './index';
 
 <Meta title="Components/Icon" />
 
@@ -27,4 +27,4 @@ import {Icon} from 'dooboo-ui';
 
 ### All Icons
 
-<Default />
+<Icon/>
