+++ docs/src/components/IconButton/IconButton.stories.mdx
@@ -1,5 +1,5 @@
-import {Meta} from '@storybook/addon-docs/blocks';
-import {Light, Dark} from './index';
+import {Meta} from '@storybook/addon-docs';
+import IconButton from './index';
 
 <Meta title="Components/IconButton" />
 
@@ -35,8 +35,4 @@ import {IconButton} from 'dooboo-ui';
 
 ## Demo
 
-#### Light Theme
-<Light />
-
-#### Dark Theme
-<Dark />
+<IconButton/>
